+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 Change log
+v1.1.1
+- Fix missing listSeedAddresses function header
+
+v1.1.0
+- Add listing seed adresses
+
+v1.0.10
+- Fix JWT signing for HD signer
+
+v1.0.9
+- Fix jwt signing to return a dictionary
 
 v1.0.8
 - Add support for deleting keys